Add ACL to newly created folders
When a new folder is created on the nuxeo server, if a file (named as specified by the acl.filename property) is present and contains an ACL rule for this folder then it is applied.
If neither file is present nor the acl.filename property is set, no ACL is set.